backend/requirements.txt
SQLAlchemy~=2.0.43
pandas~=2.2.0
requests~=2.32.5
fastapi~=0.118.0
pydantic~=2.11.9
neo4j~=6.0.2
langchain-openai~=0.3.34
langgraph~=0.6.10
langchain-core~=0.3.77
langchain-deepseek~=0.1.4
langgraph-cli[inmem]
# langgraph-supervisor 已移除，使用原生条件边实现
# langgraph-supervisor~=0.0.29
langgraph-checkpoint-postgres
langchain-mcp-adapters~=0.1.11
langchain-ollama~=0.3.10
pydantic-settings~=2.11.0
passlib[bcrypt]~=1.7.4
python-jose[cryptography]~=3.3.0
email-validator~=2.1.0
sqlparse~=0.5.3
typing_extensions~=4.15.0
uvicorn~=0.37.0
PyMySQL~=1.1.2
numpy~=1.26.0
pymilvus~=2.6.2
python-dotenv~=1.0.0
psycopg[binary,pool]
